requirements.txt
aiofiles==24.1.0
aiogram==3.18.0
aiohappyeyeballs==2.5.0
aiohttp==3.11.13
aiosignal==1.3.2
aiosqlite==0.21.0
annotated-types==0.7.0
attrs==25.1.0
certifi==2025.1.31
environs==14.1.1
frozenlist==1.5.0
idna==3.10
iniconfig==2.0.0
iso8601==2.1.0
magic-filter==1.0.12
marshmallow==3.26.1
multidict==6.1.0
packaging==24.2
pluggy==1.5.0
propcache==0.3.0
pyaes==1.6.1
pyasn1==0.6.1
pydantic==2.10.6
pydantic_core==2.27.2
pypika-tortoise==0.5.0
pytest==8.3.5
pytest-asyncio==0.25.3
python-dotenv==1.0.1
pytz==2025.1
rsa==4.9
Telethon==1.39.0
tortoise-orm==0.24.2
typing_extensions==4.12.2
yarl==1.18.3
